Remove consul for cloud
BREAKING CHANGES: Consul client is not any longer used in case of running in a cloud environment. Most `node-consul` functions are defined empty or throwing an error in cloud.
If running in cloud or on premise is configured using the environment variable `SERVICE_DISCOVERY`. Possible values are `cloud` and `consul`, default is `consul`.
Additionally the overall service port may be configured using the environment variable `SERVICE_DISCOVERY_PORT`, default is 3000.